refactor: use `feather-cored` in cmd dir and add `chain_id` to transformations
