+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "chain_id": "feather-1",
   "app_name": "feather-core",
-  "address_prefix": "pfeath",
+  "address_prefix": "feath",
   "bond_denom": "stake",
   "bond_supply": "1000000000",
   "unbonding_time": "1814400s",
